requirements.txt
bigartm==0.10.0
blis==0.4.1
catalogue==0.0.8
certifi==2019.11.28
chardet==3.0.4
cymem==2.0.3
FastNLP==0.5.0
idna==2.8
importlib-metadata==1.3.0
more-itertools==8.0.2
murmurhash==1.0.2
nltk==3.4.5
numpy==1.18.0
pkg-resources==0.0.0
plac==1.1.3
preshed==3.0.2
prettytable==0.7.2
PyYAML==5.3
requests==2.22.0
six==1.13.0
spacy==2.2.3
srsly==0.2.0
thinc==7.3.1
torch==1.3.1
tqdm==4.41.0
urllib3==1.25.7
wasabi==0.4.2
zipp==0.6.0
